+++ Source/SwiftLintFramework/Linter.swift
@@ -16,7 +16,17 @@ public struct Linter {
     private let rules: [Rule]
 
     public var styleViolations: [StyleViolation] {
-        return rules.flatMap { $0.validateFile(self.file) }
+        let regions = file.regions()
+        return rules.flatMap { rule in
+            return rule.validateFile(self.file).filter { styleViolation in
+                guard let violationRegion = regions.filter({
+                    $0.start <= styleViolation.location && $0.end >= styleViolation.location
+                }).first else {
+                    return true
+                }
+                return !violationRegion.disabledRuleIdentifiers.contains(rule.identifier)
+            }
+        }
     }
 
     public var ruleExamples: [RuleExample] {