+++ CHANGELOG.md
@@ -23,6 +23,11 @@
   names to start with an underscore.
   [JP Simard](https://github.com/jpsim)
 
+* Disable and re-enable rules from within source code comments using
+  `// swiftlint:disable $IDENTIFIER` and `// swiftlint:enable $IDENTIFIER`.  
+  [JP Simard](https://github.com/jpsim)
+  [#4](https://github.com/realm/SwiftLint/issues/4)
+
 ##### Bug Fixes
 
 * None.